added dto refactored service

+++ src/main/java/com/microservice/user/repository/SkillRepository.java
@@ -16,6 +16,6 @@ public interface SkillRepository extends JpaRepository<SkillEntity, Long> {
     /*@Query("SELECT s FROM SkillEntity s JOIN s.portfolios p JOIN p.user u WHERE u.userId = :userId")
     List<SkillEntity> findSkillsByUserId(@Param("userId") int userId);*/
     @Query("SELECT s FROM SkillEntity s JOIN s.portfolios p JOIN UserEntity u ON u.portfolio = p WHERE u.userId = :userId")
-    Optional<List<SkillEntity>> findSkillsByUserId(@Param("userId") int userId);
+    Optional<List<SkillEntity>> findSkillsByUserId(@Param("userId") Long userId);
     Optional<SkillEntity> findBySkillName(String skillName);
 }
